# public/swagger.yaml
components:
...
    schemas:
...
        Role:
            enum:
                - ADMIN
                - SHELTER
                - USER
            type: string
...
        EmailPayload:
            properties:
                email:
                    type: string
                    description: 'Email of user'
                    example: johndoe@example.com
            required:
                - email
            type: object
            additionalProperties: false
...
        Status:
            enum:
                - 'Under Review'
                - 'Home Visit Requested'
                - 'Home Visit Scheduled'
                - 'Home Approved'
                - 'Home Rejected'
                - 'User Visit Scheduled'
                - Approved
                - Rejected
                - Closed
                - Expired
                - 'Reactivation Requested'
                - 'Reactivation Request Approved'
                - 'Reactivation Request Declined'
            type: string
        ResidenceType:
            enum:
                - ownHouse
                - rentHouse
            type: string
        ApplicationResponse:
            properties:
                shelterID:
                    type: string
                microchipID:
                    type: string
                residenceType:
                    $ref: '#/components/schemas/ResidenceType'
                hasRentPetPermission:
                    type: boolean
                hasChildren:
                    type: boolean
                childrenAges:
                    type: string
                hasOtherPets:
                    type: boolean
                otherPetsInfo:
                    type: string
                petAloneTime:
                    type: number
                    format: double
                hasPlayTimeParks:
                    type: boolean
                petActivities:
                    type: string
                handlePetIssues:
                    type: string
                moveWithPet:
                    type: string
                canAffordPetsNeeds:
                    type: boolean
                canAffordPetsMedical:
                    type: boolean
                petTravelPlans:
                    type: string
                petOutlivePlans:
                    type: string
                id:
                    type: string
                status:
                    $ref: '#/components/schemas/Status'
                submissionDate:
                    type: string
                petImage:
                    type: string
                petName:
                    type: string
                shelterName:
                    type: string
                applicantName:
                    type: string
                homeVisitDate:
                    type: string
                shelterVisitDate:
                    type: string
                homeVisitEmailSentDate:
                    type: string
                shelterVisitEmailSentDate:
                    type: string
            required:
                - shelterID
                - microchipID
                - residenceType
                - hasRentPetPermission
                - hasChildren
                - childrenAges
                - hasOtherPets
                - otherPetsInfo
                - petAloneTime
                - hasPlayTimeParks
                - petActivities
                - handlePetIssues
                - moveWithPet
                - canAffordPetsNeeds
                - canAffordPetsMedical
                - petTravelPlans
                - petOutlivePlans
                - id
                - status
                - submissionDate
                - petImage
                - petName
                - shelterName
            type: object
            additionalProperties: false
        ApplictionResponseForUser:
            properties:
                application:
                    $ref: '#/components/schemas/ApplicationResponse'
                canReview:
                    type: boolean
            required:
                - application
            type: object
            additionalProperties: false
        ApplicationPayload:
            properties:
                shelterID:
                    type: string
                microchipID:
                    type: string
                residenceType:
                    $ref: '#/components/schemas/ResidenceType'
                hasRentPetPermission:
                    type: boolean
                hasChildren:
                    type: boolean
                childrenAges:
                    type: string
                hasOtherPets:
                    type: boolean
                otherPetsInfo:
                    type: string
                petAloneTime:
                    type: number
                    format: double
                hasPlayTimeParks:
                    type: boolean
                petActivities:
                    type: string
                handlePetIssues:
                    type: string
                moveWithPet:
                    type: string
                canAffordPetsNeeds:
                    type: boolean
                canAffordPetsMedical:
                    type: boolean
                petTravelPlans:
                    type: string
                petOutlivePlans:
                    type: string
            required:
                - shelterID
                - microchipID
                - residenceType
                - hasRentPetPermission
                - hasChildren
                - childrenAges
                - hasOtherPets
                - otherPetsInfo
                - petAloneTime
                - hasPlayTimeParks
                - petActivities
                - handlePetIssues
                - moveWithPet
                - canAffordPetsNeeds
                - canAffordPetsMedical
                - petTravelPlans
                - petOutlivePlans
            type: object
            additionalProperties: false
            example:
                shelterID: 6475e9630044288a2b4880b5
                microchipID: A123456789
                residenceType: rentHouse
                hasRentPetPermission: 'true'
                hasChildren: 'true'
                childrenAges: '2,4'
                hasOtherPets: 'true'
                otherPetsInfo: 'Cat 1, Dog 2'
                petAloneTime: '4'
                hasPlayTimeParks: 'true'
                petActivities: 'walking, playing around the house, swimming, sky diving'
                handlePetIssues: 'i will pat him'
                moveWithPet: 'i will take him with me'
                canAffordPetsNeeds: 'true'
                canAffordPetsMedical: 'true'
                petTravelPlans: 'i will take him wherever i go'
                petOutlivePlans: 'My friend, Alex will take care of him'
        AllApplicationsResponse:
            properties:
                applications:
                    items:
                        $ref: '#/components/schemas/ApplicationResponse'
                    type: array
                totalPages:
                    type: number
                    format: double
                applicationStatuses:
                    items:
                        type: string
                    type: array
            required:
                - applications
                - totalPages
                - applicationStatuses
            type: object
            additionalProperties: false
...
        ShelterProfileResponse:
            properties:
                profilePhoto:
                    type: string
                name:
                    type: string
                email:
                    type: string
                address:
                    type: string
                bio:
                    type: string
                rating:
                    type: number
                    format: double
                numberOfReviews:
                    type: number
                    format: double
                canReview:
                    type: boolean
            required:
                - profilePhoto
                - name
                - email
                - address
                - bio
                - rating
                - numberOfReviews
                - canReview
            type: object
            additionalProperties: false
        ShelterResponse:
            properties:
                id:
                    type: string
                name:
                    type: string
            required:
                - id
                - name
            type: object
            additionalProperties: false
        ApplictionResponseShelter:
            properties:
                shelterID:
                    type: string
                microchipID:
                    type: string
                residenceType:
                    $ref: '#/components/schemas/ResidenceType'
                hasRentPetPermission:
                    type: boolean
                hasChildren:
                    type: boolean
                childrenAges:
                    type: string
                hasOtherPets:
                    type: boolean
                otherPetsInfo:
                    type: string
                petAloneTime:
                    type: number
                    format: double
                hasPlayTimeParks:
                    type: boolean
                petActivities:
                    type: string
                handlePetIssues:
                    type: string
                moveWithPet:
                    type: string
                canAffordPetsNeeds:
                    type: boolean
                canAffordPetsMedical:
                    type: boolean
                petTravelPlans:
                    type: string
                petOutlivePlans:
                    type: string
                id:
                    type: string
                status:
                    $ref: '#/components/schemas/Status'
                applicantEmail:
                    type: string
                applicantName:
                    type: string
                submissionDate:
                    type: string
                    format: date-time
                petImage:
                    type: string
                petName:
                    type: string
            required:
                - shelterID
                - microchipID
                - residenceType
                - hasRentPetPermission
                - hasChildren
                - childrenAges
                - hasOtherPets
                - otherPetsInfo
                - petAloneTime
                - hasPlayTimeParks
                - petActivities
                - handlePetIssues
                - moveWithPet
                - canAffordPetsNeeds
                - canAffordPetsMedical
                - petTravelPlans
                - petOutlivePlans
                - id
                - status
                - applicantEmail
                - applicantName
                - submissionDate
                - petImage
                - petName
            type: object
            additionalProperties: false
        ApplictionResponseForShelter:
            properties:
                application:
                    $ref: '#/components/schemas/ApplictionResponseShelter'
            required:
                - application
            type: object
            additionalProperties: false
        VerifyInvitationResponse:
            properties:
                email:
                    type: string
                role:
                    $ref: '#/components/schemas/Role'
            required:
                - email
                - role
            type: object
            additionalProperties: false
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: express-ts
    version: 1.0.0
    license:
        name: ISC
    contact: {}
openapi: 3.0.0
paths:
...
    /application:
        post:
            operationId: ApplyForAPet
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApplictionResponseForUser'
                            examples:
                                'Example 1':
                                    value: {application: {id: 64b14bd7ba2fba2af4b5338d, status: 'Under Review', submissionDate: '2020-01-20', petImage: 'https://example.com/images/fluffy1.jpg', petName: Fluffy, shelterName: 'Animal Haven Shelter', shelterID: 6475e9630044288a2b4880b5, microchipID: A123456789, residenceType: ownHouse, hasRentPetPermission: true, hasChildren: true, childrenAges: '2,6', hasOtherPets: true, otherPetsInfo: 'Cat 1, Dog 2', petAloneTime: 6, hasPlayTimeParks: true, petActivities: 'Playing fetch, walking', handlePetIssues: 'I will buy new shoes', moveWithPet: 'I will take him with me', canAffordPetsNeeds: true, canAffordPetsMedical: true, petTravelPlans: 'I will take him wherever I go', petOutlivePlans: 'My pet will inherit my property and my friend Alex will take care of him'}, canReview: true}
            summary: 'Accepts application info and creates application if all fields are valid'
            tags:
                - Application
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApplicationPayload'
        get:
            operationId: GetApplicationDetails
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApplictionResponseForUser'
                            examples:
                                'Example 1':
                                    value: {application: {id: 64b14bd7ba2fba2af4b5338d, status: 'Under Review', submissionDate: '2020-01-20', petImage: 'https://example.com/images/fluffy1.jpg', petName: Fluffy, shelterName: 'Animal Haven Shelter', shelterID: 6475e9630044288a2b4880b5, microchipID: A123456789, residenceType: ownHouse, hasRentPetPermission: true, hasChildren: true, childrenAges: '2,6', hasOtherPets: true, otherPetsInfo: 'Cat 1, Dog 2', petAloneTime: 6, hasPlayTimeParks: true, petActivities: 'Playing fetch, walking', handlePetIssues: 'I will buy new shoes', moveWithPet: 'I will take him with me', canAffordPetsNeeds: true, canAffordPetsMedical: true, petTravelPlans: 'I will take him wherever I go', petOutlivePlans: 'My pet will inherit my property and my friend Alex will take care of him'}, canReview: true}
            summary: 'Returns an application''s details given id'
            tags:
                - Application
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: id
                    required: true
                    schema:
                        type: string
    /application/all:
        get:
            operationId: GetApplications
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AllApplicationsResponse'
                            examples:
                                'Example 1':
                                    value: {applications: [{id: 64b14bd7ba2fba2af4b5338d, status: 'Under Review', submissionDate: '2020-01-20', petImage: 'https://example.com/images/fluffy1.jpg', petName: Fluffy, shelterName: 'Animal Haven Shelter', shelterID: 6475e9630044288a2b4880b5, microchipID: A123456789, residenceType: ownHouse, hasRentPetPermission: true, hasChildren: true, childrenAges: '2,6', hasOtherPets: true, otherPetsInfo: 'Cat 1, Dog 2', petAloneTime: 6, hasPlayTimeParks: true, petActivities: 'Playing fetch, walking', handlePetIssues: 'I will buy new shoes', moveWithPet: 'I will take him with me', canAffordPetsNeeds: true, canAffordPetsMedical: true, petTravelPlans: 'I will take him wherever I go', petOutlivePlans: 'My pet will inherit my property and my friend Alex will take care of him', applicantName: 'John Doe', homeVisitDate: '2023-08-19T06:00:00Z', shelterVisitDate: '2023-09-02T04:00:00Z', homeVisitEmailSentDate: '2023-08-15T10:51:48.510Z', shelterVisitEmailSentDate: '2023-08-31T10:51:48.510Z'}, {id: 64b14bd7ba2fba2af4b5338e, status: Approved, submissionDate: '2020-02-25', petImage: 'https://example.com/images/buddy1.jpg', petName: Buddy, shelterName: 'Furry Friends Shelter', shelterID: 6475e9630044288a2b4880b6, microchipID: A987654321, residenceType: rentHouse, hasRentPetPermission: true, hasChildren: false, childrenAges: "", hasOtherPets: false, otherPetsInfo: "", petAloneTime: 4, hasPlayTimeParks: true, petActivities: 'Running, hiking', handlePetIssues: 'I will provide training', moveWithPet: 'I will take him with me', canAffordPetsNeeds: true, canAffordPetsMedical: true, petTravelPlans: 'I have a trusted pet sitter', petOutlivePlans: 'My sister will take care of him', applicantName: 'Jane Smith', homeVisitDate: '2023-08-19T06:00:00Z', shelterVisitDate: '2023-09-02T04:00:00Z', homeVisitEmailSentDate: '2023-08-15T10:51:48.510Z', shelterVisitEmailSentDate: '2023-08-31T10:51:48.510Z'}], totalPages: 10, applicationStatuses: ['Under Review', Approved, Rejected]}
            summary: 'Returns list of applications of a user'
            tags:
                - Application
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: page
                    required: true
                    schema:
                        format: double
                        type: number
                -
                    in: query
                    name: limit
                    required: true
                    schema:
                        format: double
                        type: number
                -
                    in: query
                    name: searchQuery
                    required: false
                    schema:
                        type: string
                -
                    in: query
                    name: applicationStatusFilter
                    required: false
                    schema:
                        type: string
...
    /shelter:
        get:
            operationId: GetShelter
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShelterProfileResponse'
                            examples:
                                'Example 1':
                                    value: {profilePhoto: 'https://example.com/shelterProfilePhoto.jpg', name: 'Animal Haven Shelter', email: info@animalhavenshelter.com, address: '251 Centre St, New York, NY 10013, USA', bio: 'Animal Haven is a non-profit organization that finds homes for abandoned cats and dogs throughout the Tri-State area, and provides behavior intervention when needed to improve chances of adoption. Founded in 1967, we operate an animal shelter in Manhattan. We also provide programs that enhance the bond between animals and people.', rating: 4.8, numberOfReviews: 567, canReview: true}
            summary: 'Returns a shelter''s details given id'
            tags:
                - Shelter
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: id
                    required: true
                    schema:
                        type: string
    /shelter/all:
        get:
            operationId: GetShelters
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/ShelterResponse'
                                type: array
                            examples:
                                'Example 1':
                                    value: {id: 60df5df17f4c6a01ac63491c, name: 'Furry Friends Sanctuary'}
            summary: 'Returns ids and names of all shelters'
            tags:
                - Shelter
            security:
                -
                    bearerAuth: []
            parameters: []
    /shelter/application:
        get:
            operationId: GetApplicationDetails
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApplictionResponseForShelter'
            summary: 'Returns an application''s details given id'
            tags:
                - Shelter
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: id
                    required: true
                    schema:
                        type: string
    /shelter/invite:
        post:
            operationId: InviteShelter
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
                            examples:
                                'Example 1':
                                    value: {email: johndoe@example.com}
            summary: 'Sends invite to shelter to sign up'
            tags:
                - Shelter
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmailPayload'
    /shelter/verifyInvitationToken:
        get:
            operationId: VerifyInvitationToken
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/VerifyInvitationResponse'
                            examples:
                                'Example 1':
                                    value: {email: info@animalhavenshelter.com, role: SHELTER}
            summary: 'Verifies the invitation token for shelter'
            tags:
                - Shelter
            security: []
            parameters:
                -
                    in: query
                    name: invitationToken
                    required: true
                    schema:
                        type: string
...
servers:
    -
        url: /
...